pyproject.toml
[project]
dependencies = ["Pillow", "ascii_magic"]
name = "terminal-flag-banner"
version = "0.1.1"
description = "A Python package to display a banner of flags in the terminal."
authors = [{ name = "Joseph Kelly", email = "jsph.kelly@outlook.com" }]
readme = "README.md"
license = { text = "MIT License" }
requires-python = ">=3.10"

[project.scripts]
terminal-flag-banner = "terminal_flag_banner.__main__:main"

[tool.setuptools.package-data]
terminal-flag-banner = ["data/*"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
